requirements.txt
# LinkedIn Photo Scraper v3 — dépendances
# pip install -r requirements.txt

streamlit>=1.30.0
patchright>=1.58.0
httpx>=0.27.0
face_recognition>=1.3.0
nest_asyncio>=1.6.0
Pillow>=10.0.0
numpy>=1.24.0
